Finished 9_2

+++ src/day_9/utils.rs
@@ -1,7 +1,7 @@
 use std::collections::VecDeque;
 
 /// Parses the game rules text to internal data representation
-pub fn get_game_rules(rules_text: &str) -> (usize, usize) {
+pub fn get_game_rules(rules_text: &str) -> (u64, u64) {
     let regex = regex::Regex::new(r"([0-9])\w+").unwrap();
     let mut matches = regex.find_iter(rules_text);
     let num_players = matches.next().unwrap().as_str().parse().unwrap();
@@ -9,10 +9,10 @@ pub fn get_game_rules(rules_text: &str) -> (usize, usize) {
     (num_players, num_marbles)
 }
 
-pub fn play_marble_game(num_players: usize, num_marbles: usize) -> usize {
-    let mut circle = VecDeque::with_capacity(num_marbles);
+pub fn play_marble_game(num_players: u64, num_marbles: u64) -> u64 {
+    let mut circle = VecDeque::with_capacity(num_marbles as usize);
     circle.push_front(0);
-    let mut scores = vec![0_usize; num_players];
+    let mut scores = vec![0_u64; num_players as usize];
 
     for marble in 1..num_marbles {
         if marble % 23 == 0 {
@@ -20,7 +20,7 @@ pub fn play_marble_game(num_players: usize, num_marbles: usize) -> usize {
             let score_m = marble;
             let score_p = circle.pop_front().unwrap();
             let score = score_m + score_p;
-            scores[marble % num_players] += score;
+            scores[(marble % num_players) as usize] += score;
         } else {
             rotate(&mut circle, -2);
             circle.push_front(marble);
@@ -37,7 +37,7 @@ fn rotate<T>(deq: &mut VecDeque<T>, n: isize) {
         return;
     }
 
-    let abs = n.abs() as usize % deq.len();
+    let abs = n.abs() as u64 % deq.len() as u64;
     if n >= 0 {
         for _ in 0..abs {
             let popped = deq.pop_back();
